end_to_end_tests/golden-record/pyproject.toml
[tool.poetry]
name = "my-test-api-client"
version = "0.1.0"
description = "A client library for accessing My Test API"
authors = []
readme = "README.md"
packages = [
    {include = "my_test_api_client"},
]
include = ["CHANGELOG.md", "my_test_api_client/py.typed"]


[tool.poetry.dependencies]
python = "^3.9"
httpx = ">=0.20.0,<0.28.0"
attrs = ">=21.3.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
